requirements.txt
# PyPalette Requirements
# ====================
# Required packages for PyPalette color palette tool

# Core dependencies
numpy>=1.21.0
Pillow>=8.3.0

# GUI application
PyQt5>=5.15.0

# Build tools
pyinstaller>=4.5.0

# Optional dependencies for enhanced functionality
matplotlib>=3.5.0  # For advanced color visualizations (optional)
opencv-python>=4.5.0  # For advanced image processing (optional)

# Development dependencies (uncomment if developing)
# pytest>=6.0.0
# black>=21.0.0
# flake8>=3.9.0
# mypy>=0.910
